Move otp2 types to their own subdirectory within types package
